TUI implemented

+++ modules/station.py
@@ -48,26 +48,27 @@ class control:
                 p = audio.effect.normalize(self.db, file)
                 if p != None:
                     procs.append((p, file))
+        
+        if len(procs) > 0:
+            logging.info("Waiting for loudness normalization...")
 
-        logging.info("Waiting for loudness normalization...")
+            try:
+                for p in procs:
+                    p[0].join()
 
-        try:
-            for p in procs:
-                p[0].join()
+                while len(multiprocessing.active_children()) > 0:
+                    time.sleep(1)
 
-            while len(multiprocessing.active_children()) > 0:
-                time.sleep(1)
-
-            for p in procs:  # t = (thread, file)
-                p[0].close()
-                file = p[1]
-                db.setRecord(self.db, file, fsUtil.sha256sum(
-                    file), LOUDNESS, BITRATE)
-                procs.remove(p)
-        except Exception as e:
-            logging.error("Process error: " + str(e))
-            for p in multiprocessing.active_children():
-                p.kill()
+                for p in procs:  # t = (thread, file)
+                    p[0].close()
+                    file = p[1]
+                    db.setRecord(self.db, file, fsUtil.sha256sum(
+                        file), LOUDNESS, BITRATE)
+                    procs.remove(p)
+            except Exception as e:
+                logging.error("Process error: " + str(e))
+                for p in multiprocessing.active_children():
+                    p.kill()
 
         logging.info("All sounds in lib normalized.")
